library/requirements.txt
ansible==9.4.0
basyx-python-sdk==1.0.0